More wasm backend triage
