+++ tests/core/test_emscripten_api.cpp
