+++ tests/test_pystruct.out
@@ -1,0 +1,7 @@
+*0,0,0,4,8,16,20,24*
+*1,0,0*
+*0*
+0:1,1
+1:1,1
+2:1,1
+*16,24,24*
